move cpu metrics to dedicated folder
